moves postprocess_rois to modules
