+++ src/ophys_etl/pipelines/segment_postprocess_pipeline.py
@@ -5,8 +5,9 @@ import tempfile
 import marshmallow
 import argschema
 
-from ophys_etl.transforms.postprocess_rois import (
-    PostProcessROIsInputSchema, PostProcessROIs)
+from ophys_etl.modules.postprocess_rois.__main__ import PostProcessROIs
+from ophys_etl.modules.postprocess_rois.schemas import \
+        PostProcessROIsInputSchema
 from ophys_etl.modules.suite2p_wrapper.__main__ import Suite2PWrapper
 from ophys_etl.modules.suite2p_wrapper.schemas import Suite2PWrapperSchema
